% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dscoretest.R
\name{debias_standard}
\alias{debias_standard}
\title{Standard debiasing}
\usage{
debias_standard(
  h.hat,
  X.debias,
  fit.debias,
  predict_fun,
  weight_fun,
  wls_method,
  arg.wls_method
)
}
\arguments{
\item{h.hat}{A list as returned by one of \code{\link[=hunt_optimal]{hunt_optimal()}}, \code{\link[=hunt_wls]{hunt_wls()}}, \code{\link[=hunt_vanilla]{hunt_vanilla()}}.}

\item{X.debias}{Part of X for debiasing.}

\item{fit.debias}{Null model fitted on the debiasing sample of X and y.}

\item{predict_fun, weight_fun, wls_method, arg.wls_method}{They must be compatible with
\code{fit.debias}; see \code{\link[=dScoreTest]{dScoreTest()}} for details.}
}
\value{
A list with elements:
\describe{
\item{\code{m.h.fit}}{The null model fitted (over all columns of \code{X})
to project and debias \eqn{\hat{h}}.}
\item{\code{h}}{The debiased hunt function \eqn{\hat{h} - \hat{m}_{\hat{h}}}
with signature \code{h(X)}.}
}
}
\description{
Transform a hunted function \eqn{\hat{h}} into a debiased function
\eqn{\hat{h} - \hat{m}_{\hat{h}}}, where \eqn{\hat{m}_{\hat{h}}} is the
projection of \eqn{\hat{h}} onto the null model.
}
\details{
The projection \eqn{\hat{m}_{\hat{h}}} is obtained by fitting the null model
(via \code{wls_method}, weighted by \code{weight_fun(fit.debias, X.debias)})
with the hunted values \code{h.hat$h(X.debias)} as response. This projection
uses \strong{all} columns of \code{X}, even when the hunt itself is driven by
only a subset of covariates (\code{h.hat$X.cols}).
}
