% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dscoretest.R
\name{plot.dScoreTest}
\alias{plot.dScoreTest}
\title{Plot the score test}
\usage{
\method{plot}{dScoreTest}(x, ...)
}
\description{
Diagonotic plots for the test:
\enumerate{
\item Distribution of \eqn{\{L_i\}}. Extremes values under the null can result
in bad normal approximation.
In this case, consider set \code{trim.outlier.hunt=TRUE}.
\item Plots on debiasing. Left: resids (nagetive scores) versus the hunted
signal. A horizontal segment is drawn between each pair of raw hunted
signal and the debiased hunted signal. If debiased gets higher, colored
in red; otherwise colored in green. Right: L = resids x hunt before and
after debiasing.
}
}
