% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dscoretest.R
\name{print.dScoreTest}
\alias{print.dScoreTest}
\title{Print the score test}
\usage{
\method{print}{dScoreTest}(x, ...)
}
\arguments{
\item{x}{A \code{dScoreTest} object.}

\item{...}{Unused, for S3 consistency.}
}
\description{
Print the score test
}
