% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova.R
\name{compare_models}
\alias{compare_models}
\title{Compare two models using the debiased score test}
\usage{
compare_models(fit.0, ...)
}
\arguments{
\item{fit.0}{A fitted null-model object. Methods are provided for
\code{glm}, \code{lm} and \code{mgcv::gam} fits.}

\item{...}{Additional arguments passed to the dispatched method,
notably the alternative-model fit \code{fit.1}.}
}
\description{
Compare two models using the debiased score test
}
\seealso{
\code{\link{compare_models.glm}}, \code{\link{compare_models.lm}},
\code{\link{compare_models.gam}},
\code{\link{gof_test}}, \code{\link{dScoreTest}}
}
