% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dScoreTest-package.R, R/dscoretest.R
\docType{package}
\name{dScoreTest}
\alias{dScoreTest-package}
\alias{dScoreTest}
\title{Debiased score test: goodness-of-fit test and model comparison}
\usage{
dScoreTest(
  y,
  X,
  score_fun,
  weight_fun,
  fit_method,
  wls_method,
  hunt.style = "optimal",
  hunt.method = "grf",
  hunt_fun = NULL,
  trim.outlier.hunt = TRUE,
  X.cols.hunt = 1:ncol(X),
  splits = c(0.5, 0.5),
  arg.fit_method = NULL,
  arg.wls_method = NULL,
  arg.hunt_fun = NULL,
  predict_fun = stats::predict,
  predict_fun_alt = NULL,
  verbose = FALSE
)
}
\arguments{
\item{y}{Numeric response vector of length n.}

\item{X}{Numeric covariate matrix of dimension n x p.}

\item{score_fun}{Function with signature \code{score_fun(fit, y, X)}
returning a vector of scores \eqn{l'(\hat{f}(x_i), y_i)}, which can be viewed
as negative residuals.}

\item{weight_fun}{Function with signature \code{weight_fun(fit, X)}
that computes the weight \eqn{\mathbb{E}[l''(\hat{f}(x_i), y_i) | x_i]} for each
row \eqn{x_i} of X.}

\item{fit_method}{Function with signature \code{fit_method(y, X, ...)}
that returns a fitted null model \eqn{\hat{f} \in \mathcal{F}} by
minimizing the loss \eqn{\sum_i l(f(x_i), y_i)}. For a fitted \code{f},
it must support \code{predict_fun(f, X)} for evaluation.}

\item{wls_method}{Function with signature \code{wls_method(y, X, w, ...)}
that fits the null model \eqn{\hat{f} \in \mathcal{F}} with weighted least
squares, i.e., minimizing \eqn{\sum_i w_i (f(x_i) - y_i)^2}. For a fitted
\code{f}, it must support \code{predict_fun(f, X)} for evaluation.}

\item{hunt.style}{Hunting algorithm with the following options.
\itemize{
\item \code{'optimal'}: optimal hunting (default).
See \code{\link{hunt_optimal}}.
\item \code{'wls'}: a simpler hunting using weighted least squares,
which can be less powerful. See \code{\link{hunt_wls}}.
\item \code{'vanilla'}: a basic hunting; not
recommended unless unable to fit an alternative model with weighted
least squares.  See \code{\link{hunt_vanilla}}.}}

\item{hunt.method}{Built-in method for hunting. Currently available:
\itemize{
\item \code{'grf'}: regression forest from package \code{grf}.
}
When this is set to any other value, arguments \code{hunt_fun} and
\code{predict_fun_alt} must be set properly to supply a customized
hunting method.}

\item{hunt_fun}{Default \code{NULL}.
When \code{hunt.method} is not set to a built-in method,
this is a customized function for hunting. When \code{hunt.style} is
\code{'optimal'} or \code{'wls'}, this function must have signature
\code{hunt_fun(y, X, w, ...)} that returns a fitted
\emph{alternative model} \eqn{\hat{g} \in \mathcal{G}} via weighted least
squares, i.e., by minimizing \eqn{\sum_i w_i (y_i - g(x_i))^2};
otherwise, for \code{'vanilla'} hunting,
this function must have signature \code{hunt_fun(y, X, ...)} that
returns an \emph{alternative model} fitted in any fashion.
The returned object \code{g} must support \code{predict_fun_alt(g, X)}
for evaluation.}

\item{trim.outlier.hunt}{If \code{TRUE} (default),
extreme values produced by the hunted function will be trimmed using Tukey's
IQR rule.}

\item{X.cols.hunt}{Integer vector selecting which columns of
\code{X} drive the hunt. Default \code{1:ncol(X)}. This is modified only
in special settings, e.g., when there is an offset in the null model.}

\item{splits}{Numeric vector of length 2 or 3 giving the relative sizes
of the sample splits; rescaled internally to sum to one.
Default is \code{c(0.5, 0.5)}, which splits data into two halves for
hunt and test respectively. Though typically unnecessary in practice,
one can also specify a 3-way split for hunt, debiasing and test respectively.}

\item{arg.fit_method}{Named list of additional arguments passed to
\code{fit_method} (default to \code{NULL}).}

\item{arg.wls_method}{Named list of additional arguments passed to
\code{wls_method} (default to \code{NULL}).}

\item{arg.hunt_fun}{Extra arguments (default \code{NULL}) passed to the
customized \code{hunt.fun}.}

\item{predict_fun}{Function with signature \code{predict_fun(fit, X)}
returning a numeric vector of predictions from a fitted null model, which
is produced by \code{fit_method()} and \code{wls_method()}.
Note that if \code{fit} is \eqn{\hat{f}}, this function should return
\eqn{\hat{f}(X)}. Default \code{stats::predict}.
When y is binary, it must also support signature
\code{predict_fun(fit, X, type='response')} for returning probabilities.}

\item{predict_fun_alt}{Default \code{NULL}.
When \code{hunt.method} is not set to a built-in method, this is a function
with signature \code{predict_fun_alt(fit, X)}
returning a numeric vector of predictions from a fitted alternative model
produced by \code{hunt_fun()}.}

\item{verbose}{Default \code{FALSE}; information is printed if set to
\code{TRUE}.}
}
\value{
An object of class \code{"dScoreTest"}.
}
\description{
Test whether a parametric (e.g., glm) or a semiparametric (e.g., GAM) model
is well-specified. The test is a debiased (Neyman-orthogonalized) score test
computed via sample splitting:
on a held-out hunt sample, a flexible auxiliary fit is hunted for a
direction in which the null model's score is non-zero; on a held-out test
sample, that direction's score is evaluated and standardized under the
null. The orthogonalization absorbs plug-in bias from estimating the
direction, so the resulting test statistic is asymptotically standard
normal under the null without requiring a parametric form for the
alternative.
}
\details{
For most scenarios, use one of these methods instead:

\itemize{
\item Use \code{\link{gof_test}} to test whether a fitted model is
well-specified against a nonparametric alternative. S3 methods are
provided for \code{glm} (\code{\link{gof_test.glm}}), \code{lm}
(\code{\link{gof_test.lm}}) and \code{mgcv::gam}
(\code{\link{gof_test.gam}}).

\item Use \code{\link{compare_models}} to test a null model \code{fit.0}
against an alternative supermodel \code{fit.1} in the same model class.
Similar to \code{\link{stats:anova}}, method can be used to conduct a
significance test of one or more predictors.
In contrast with \code{\link{gof_test}}, this method targets the alternative
\code{fit.1}.
S3 methods are provided for \code{glm} (\code{\link{compare_models.glm}}),
\code{lm} (\code{\link{compare_models.lm}}) and \code{mgcv::gam}
(\code{\link{compare_models.gam}}).}

Use \code{\link{dScoreTest}} directly for full control over the score,
weight, refit and hunt routines: this is the underlying engine that the
S3 methods wrap.
}
\seealso{
\code{\link{hunt_optimal}}, \code{\link{hunt_wls}},
\code{\link{hunt_vanilla}}, \code{\link{new_dScoreTest}}
}
\author{
\strong{Maintainer}: Richard Guo \email{ricguo@umich.edu} (\href{https://orcid.org/0000-0002-2081-7398}{ORCID})

Authors:
\itemize{
  \item Aditya Dhawan \email{ad950@cam.ac.uk}
}

}
