% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gof.R
\name{gof_test.gam}
\alias{gof_test.gam}
\title{Goodness-of-fit test for a GAM}
\usage{
\method{gof_test}{gam}(
  object,
  hunt.style = "optimal",
  hunt.method = "grf",
  hunt_fun = NULL,
  trim.outlier.hunt = TRUE,
  X.cols.exclude = NULL,
  splits = c(0.5, 0.5),
  arg.hunt_fun = NULL,
  predict_fun_alt = NULL,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{object}{Fitted \code{mgcv::gam} object.}

\item{hunt.style}{Hunting algorithm with the following options.
\itemize{
\item \code{'optimal'}: optimal hunting (default).
See \code{\link{hunt_optimal}}.
\item \code{'wls'}: a simpler hunting using weighted least squares,
which can be less powerful. See \code{\link{hunt_wls}}.
\item \code{'vanilla'}: a basic hunting; not
recommended unless unable to fit an alternative model with weighted
least squares.  See \code{\link{hunt_vanilla}}.}}

\item{hunt.method}{Built-in method for hunting. Currently available:
\itemize{
\item \code{'grf'}: regression forest from package \code{grf}.
}
When this is set to any other value, arguments \code{hunt_fun},
\code{arg.hunt_fun} and \code{predict_fun_alt} are used to specify a
customized hunting method.}

\item{hunt_fun}{Default \code{NULL}.
When \code{hunt.method} is not set to a built-in method,
this is a customized function for hunting. When \code{hunt.style} is
\code{'optimal'} or \code{'wls'}, this function must have signature
\code{hunt_fun(y, X, w, ...)} that returns a fitted
\emph{alternative model} \eqn{\hat{g} \in \mathcal{G}} via weighted least
squares, i.e., by minimizing \eqn{\sum_i w_i (y_i - g(x_i))^2};
otherwise, for \code{'vanilla'} hunting,
this function must have signature \code{hunt_fun(y, X, ...)} that
returns an \emph{alternative model} fitted in any fashion.
The returned object \code{g} must support \code{predict_fun_alt(g, X)}
for evaluation.}

\item{trim.outlier.hunt}{If \code{TRUE} (default),
extreme values produced by the hunted function will be trimmed using Tukey's
IQR rule.}

\item{X.cols.exclude}{Columns in \code{stats::model.matrix(object)} to be
excluded when hunting for alternative signal. Default \code{NULL}.}

\item{splits}{Numeric vector of length 2 or 3 giving the relative sizes
of the sample splits; rescaled internally to sum to one.
Default is \code{c(0.5, 0.5)}, which splits data into two halves for
hunt and test respectively. Though typically unnecessary in practice,
one can also specify a 3-way split for hunt, debiasing and test respectively.}

\item{arg.hunt_fun}{Extra arguments (default \code{NULL}) passed to the
customized \code{hunt.fun}.}

\item{predict_fun_alt}{When a customized \code{hunt.fun} is used, this is a
function with signature \code{predict_fun_alt(fit, X)} returning a numeric
vector of predictions from a fitted alternative model produced by
\code{hunt_fun()}.}

\item{verbose}{Default \code{FALSE}; information is printed if set to
\code{TRUE}.}

\item{...}{Unused; present for S3 generic/method consistency.}
}
\value{
An object of class \code{"dScoreTest"}: a list whose key elements
are the debiased test statistic \code{t.stat} and the one-sided p-value
\code{p.val} (right tail of the standard normal), along with the test-set
score residuals, the hunted direction, and the call. It has
\code{\link[=print.dScoreTest]{print}},
\code{\link[=summary.dScoreTest]{summary}} and
\code{\link[=plot.dScoreTest]{plot}} methods.
}
\description{
Debiased score test for goodness of fit of an \code{mgcv::gam} fit.
}
\details{
Only the numeric predictors appearing in \code{stats::model.frame(object)}
are exposed to the hunt; \code{X.cols.exclude} indexes into these
predictor variables (not basis columns). Factor-by smooths and other
non-numeric predictors are not currently supported. Formulas using
\code{offset()} terms, a \code{weights} argument, or a multi-column
response (e.g. \code{cbind(succ, fail) ~ ...}) are also not supported.
}
\examples{
 set.seed(42)
 dat <- mgcv::gamSim(eg=1, n=400, dist="normal", scale=2, verbose = FALSE)
 dat.0 <- dat[,1:5]
 
 # well-specified
 fit.0 <- mgcv::gam(y~s(x0)+s(x1)+s(x2)+s(x3),data=dat.0)
 test.0 <- gof_test(fit.0)
 # f3=0, also well-specified
 \donttest{fit.1 <- mgcv::gam(y~s(x0)+s(x1)+s(x2),data=dat.0)
 test.1 <- gof_test(fit.1)
 plot(test.1)}
 # misspecified
 \donttest{dat.1 <- dat.0
 dat.1$y <- dat.1$y * dat$f0 
 fit.2 <- mgcv::gam(y~s(x0)+s(x1)+s(x2)+s(x3), data=dat.1)
 test.2 <- gof_test(fit.2)
 plot(test.2)}

}
