% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gof.R
\name{gof_test.glm}
\alias{gof_test.glm}
\title{Goodness-of-fit test for GLM}
\usage{
\method{gof_test}{glm}(
  object,
  hunt.style = "optimal",
  hunt.method = "grf",
  hunt_fun = NULL,
  trim.outlier.hunt = TRUE,
  X.cols.exclude = NULL,
  splits = c(0.5, 0.5),
  arg.hunt_fun = NULL,
  predict_fun_hunt = NULL,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{object}{Fitted glm object.}

\item{hunt.style}{Hunting algorithm with the following options.
\itemize{
\item \code{'optimal'}: optimal hunting (default).
See \code{\link{hunt_optimal}}.
\item \code{'wls'}: a simpler hunting using weighted least squares,
which can be less powerful. See \code{\link{hunt_wls}}.
\item \code{'vanilla'}: a basic hunting; not
recommended unless unable to fit an alternative model with weighted
least squares.  See \code{\link{hunt_vanilla}}.}}

\item{hunt.method}{Built-in method for hunting. Currently available:
\itemize{
\item \code{'grf'}: regression forest from package \code{grf}.
}
When this is set to any other value, arguments \code{hunt_fun},
\code{arg.hunt_fun} and \code{predict_fun_hunt} are used to specify a
customized hunting method.}

\item{hunt_fun}{Default \code{NULL}.
When \code{hunt.method} is not set to a built-in method,
this is a customized function for hunting. When \code{hunt.style} is
\code{'optimal'} or \code{'wls'}, this function must have signature
\code{hunt_fun(y, X, w, ...)} that returns a fitted
\emph{alternative model} \eqn{\hat{g} \in \mathcal{G}} via weighted least
squares, i.e., by minimizing \eqn{\sum_i w_i (y_i - g(x_i))^2};
otherwise, for \code{'vanilla'} hunting,
this function must have signature \code{hunt_fun(y, X, ...)} that
returns an \emph{alternative model} fitted in any fashion.
The returned object \code{g} must support \code{predict_fun_hunt(g, X)}
for evaluation.}

\item{trim.outlier.hunt}{If \code{TRUE} (default),
extreme values produced by the hunted function will be trimmed using Tukey's
IQR rule.}

\item{X.cols.exclude}{Columns in \code{stats::model.matrix(object)} to be
excluded when hunting for alternative signal. Default \code{NULL}.}

\item{splits}{Numeric vector of length 2 or 3 giving the relative sizes
of the sample splits; rescaled internally to sum to one.
Default is \code{c(0.5, 0.5)}, which splits data into two halves for
hunt and test respectively. Though typically unnecessary in practice,
one can also specify a 3-way split for hunt, debiasing and test respectively.}

\item{arg.hunt_fun}{Extra arguments (default \code{NULL}) passed to the
customized \code{hunt.fun}.}

\item{predict_fun_hunt}{When a customized \code{hunt.fun} is used, this is a
function with signature \code{predict_fun_hunt(fit, X)} returning a numeric
vector of predictions from a fitted alternative model produced by
\code{hunt_fun()}.}

\item{verbose}{Default \code{FALSE}; information is printed if set to
\code{TRUE}.}

\item{...}{Unused; present for S3 generic/method consistency.}
}
\value{
An object of class \code{"dScoreTest"}: a list whose key elements
are the debiased test statistic \code{t.stat} and the one-sided p-value
\code{p.val} (right tail of the standard normal), along with the test-set
score residuals, the hunted direction, and the call. It has
\code{\link[=print.dScoreTest]{print}},
\code{\link[=summary.dScoreTest]{summary}} and
\code{\link[=plot.dScoreTest]{plot}} methods.
}
\description{
Debiased score test for goodness of fit of GLM.
}
\examples{
 set.seed(42)
 n <- 500
 X <- matrix(rnorm(n * 3), nrow = n)
 # log(E[y]) ~ X well-specified
 y0 <- 5 * exp(X[,1] + X[,3]) + rnorm(n) * 3
 fit.0 <- glm(y0 ~ X, family = gaussian(link = "log"), start=rep(1,4))
 gof_test(fit.0)
 # log(E[y]) ~ X misspecified
 y1 <- y0 + exp(6 * cos(X[,1]/6)^2) / sqrt(n)
 fit.1 <- glm(y1 ~ X, family = gaussian(link = "log"), start=rep(1,4))
 gof_test(fit.1)

}
