% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gof.R
\name{gof_test}
\alias{gof_test}
\title{Debiased score test for goodness of fit}
\usage{
gof_test(object, ...)
}
\description{
Debiased score test for goodness of fit
}
\seealso{
\code{\link{gof_test.glm}}, \code{\link{gof_test.lm}},
\code{\link{gof_test.gam}},
\code{\link{compare_models}}, \code{\link{dScoreTest}}
}
