% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hunt.R
\name{hunt_wls}
\alias{hunt_wls}
\title{Weighted-least-squares hunting}
\usage{
hunt_wls(
  wls_alt_method,
  resids,
  X,
  X.cols = 1:ncol(X),
  trim.outlier = TRUE,
  arg.wls_alt_method = NULL,
  predict_fun_alt = stats::predict
)
}
\arguments{
\item{wls_alt_method}{Function with signature \code{wls_alt_method(y, X, w, ...)} that
returns a fitted \emph{alternative model} \eqn{\hat{g} \in \mathcal{G}} by
minimizing \eqn{\sum_i w_i (y_i - g(x_i))^2}. The returned object must
support \code{predict_fun_alt(g, X)} for evaluation.}

\item{resids}{Residuals (i.e., negative scores) of length n from the null model.}

\item{X}{Covariates of dim n x p.}

\item{X.cols}{Subset of covariates to hunt. (Default: \code{1:ncol(X)})}

\item{trim.outlier}{If \code{TRUE}, outliers in \eqn{\hat{h}(X)} will be
trimmed from the hunted \eqn{\hat{h}} using Tukey's IQR rule.}

\item{arg.wls_alt_method}{Named list of additional arguments passed to
\code{wls_alt_method} (default to \code{NULL}).}

\item{predict_fun_alt}{Function with signature \code{predict_fun_alt(fit, X)}
returning a numeric vector of predictions from the alternative-model fit.
Default \code{stats::predict}.}
}
\value{
A function h with signature \code{h(X)}.
}
\description{
Returns the hunted signal h as a function
}
