% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dscoretest.R
\name{plot.dScoreTest}
\alias{plot.dScoreTest}
\title{Plot the score test}
\usage{
\method{plot}{dScoreTest}(x, ...)
}
\arguments{
\item{x}{A \code{dScoreTest} object.}

\item{...}{Further graphical parameters passed to underlying plotting
functions.}
}
\value{
No return value; called for its side effect of producing the
diagnostic plots described above.
}
\description{
Diagonotic plots for the test:
\enumerate{
\item Histogram of \eqn{\{L_i\}}, where \eqn{L_i = resid_i \times h_i}.

\item \eqn{\{L_i\}} against the index \eqn{i}, where \eqn{i} refers to the
\eqn{i}-th observation in the full dataset. Only those \eqn{i}'s in the
test split are drawn. The mean is drawn as a horizontal line.
Extremes values under the null can result in bad normal approximation.
In this case, consider setting \code{trim.outlier.hunt=TRUE}.

\item Residuals (negative scores) versus the hunted signal.
A horizontal segment is drawn between each pair of raw hunted
signal and the debiased hunted signal. If debiased gets higher, colored
in red; otherwise colored in green. A regression line (blue) with a large
positive slope indicates the model is misspecified.

\item Normalized \eqn{\{L_i\}} drawn in order.
}
}
